requirements.txt
allure-pytest==2.8.29
allure-python-commons==2.8.29
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
enum34==1.1.10
idna==2.10
iniconfig==1.1.1
lxml==4.6.2
namedlist==1.8
packaging==20.8
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
toml==0.10.2
urllib3==1.26.2
webdriver-manager==3.2.2
